pyproject.toml
[tool.poetry]
name = "api-python-ia"
version = "0.1.0"
description = "API para MVP da Auditeste que fara a interacao entre backend em C# e o ChatGPT"
authors = ["giuanm <giuanm@live.com>"]
readme = "README.md"
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
pydantic = "^2.10.5"
fastapi = "^0.115.7"
uvicorn = "^0.34.0"
openai = "^1.60.0"
google-generativeai = "^0.8.4"
protobuf = "^5.29.3"
tenacity = "^9.0.0"
pika = "^1.3.2"
sqlalchemy = "^2.0.37"
logging = "^0.4.9.6"
uuid = "^1.30"
celery = "^5.4.0"
flower = "^2.0.1"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
